backend/calix/requirements.txt
alembic==1.18.3
asgiref==3.11.1
attrs==25.4.0
blinker==1.9.0
click==8.3.1
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-Parameter-Validation==2.3.1
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
gunicorn==25.0.3
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
packaging==26.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.5
